backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.6.0
pymongo==4.9.1
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.20
beanie==1.27.0
email-validator==2.2.0
